Merging 'gnmi' branch (0.8.1) to 'master'.

+++ test/dependencies_trusty.sh
@@ -37,14 +37,27 @@ function install_dependencies {
 }
 
 function install_gcc5 {
-    print_msg "Installing gcc-5 and g++5"
+  which gcc
+  local status=$?
+  if [[ $status == 0 ]]
+  then
+    gcc_version=$(echo `gcc --version` | awk '{ print $3 }' | cut -d '-' -f 1)
+    print_msg "Current gcc/g++ version is $gcc_version"
+  else
+    print_msg "The gcc/g++ not installed"
+    gcc_version="4.0"
+  fi
+  if [[ $(echo $gcc_version | cut -d '.' -f 1) < 5 ]]
+  then
+    print_msg "Upgrading gcc/g++ to version 5"
     sudo add-apt-repository ppa:ubuntu-toolchain-r/test -y
     sudo apt-get update > /dev/null
     sudo apt-get install gcc-5 g++-5 -y > /dev/null
-    sudo ln -f -s /usr/bin/g++-5 /usr/bin/c++
-    sudo ln -f -s /usr/bin/gcc-5 /usr/bin/cc
-    sudo ln -f -s /usr/bin/g++-5 /usr/bin/g++
-    sudo ln -f -s /usr/bin/gcc-5 /usr/bin/gcc
+    sudo ln -fs /usr/bin/g++-5 /usr/bin/c++
+    sudo ln -fs /usr/bin/gcc-5 /usr/bin/cc
+    gcc_version=$(echo `gcc --version` | awk '{ print $3 }' | cut -d '-' -f 1)
+    print_msg "Installed gcc/g++ version is $gcc_version"
+  fi
 }
 
 function install_go {
